+++ backend/models/tasks.model.js
@@ -1,6 +1,7 @@
 import { config } from "dotenv";
-import { pool } from "../service/connection.js";
-import { insertDataInTable, selectDataInTable, updateDataInTable } from "../service/duplicatePartsCode.js";
+import { pool } from "#root/service/connection.js";
+import { insertDataInTable, selectDataInTable, updateDataInTable } from "#root/service/duplicatePartsCode.js";
+import { publishMessage } from "#rmq/publisher.js";
 
 config();
 
@@ -82,28 +83,54 @@ export const TasksModel = {
       if (sql.type == "Error") {
         throw new Error(sql.message);
       }
-      const resultCreateTask = await pool.query(sql.message, sql.values);
+      const resultCreateTask = await pool.query(sql.message, Array.from(sql.values.values()));
       if (!resultCreateTask.rows.length) {
         throw new Error("The task has not been created");
       }
       if (data.hasOwnProperty("assignments")) {
-        const taskId = resultCreateTask.rows[0].task_id;
-        const usersId = data.assignments;
-        const checkUsersExists = await pool.query("SELECT user_id FROM users WHERE user_id = ANY($1)", [usersId]);
+        const task_id = resultCreateTask.rows[0].task_id;
+        const users_id = data.assignments;
+        const checkUsersExists = await pool.query("SELECT user_id FROM users WHERE user_id = ANY($1)", [users_id]);
         const existingUsersId = checkUsersExists.rows.map(row => row.user_id);
-        const nonExistingUsers = usersId.filter(userId => !existingUsersId.includes(userId));
+        const nonExistingUsers = users_id.filter(userId => !existingUsersId.includes(userId));
         if (nonExistingUsers.length > 0) {
           throw new Error(`Users with ids ${nonExistingUsers.join(", ")} do not exist`);
         }
-        for (let userId of existingUsersId) {
+        for (const user_id of existingUsersId) {
           const resultTaskAssignments = await pool.query(`INSERT INTO task_assignments (task_id, user_id, assigned_at, is_completed) \
-          VALUES ($1, $2, DEFAULT, DEFAULT) RERURNING task_assignment_id`, [taskId, userId]);
+          VALUES ($1, $2, DEFAULT, DEFAULT) RERURNING task_assignment_id`, [task_id, user_id]);
+          const notificationData = {
+            data: {
+              userId: user_id,
+              eventType: "task.assigned",
+              title: "Новая задача",
+              body: "Поручено новая задача",
+              data: {
+                task_id: task_id,
+                task_name: data.task.task_name,
+              }
+            }
+          }
+          publishMessage("task", "task.assigned", notificationData);
           if (!resultTaskAssignments.rows.length) {
             console.error("Error when insert data to task_assignment_id table", resultTaskAssignments);
             throw new Error("Error when insert data to task_assignment_id table");
           }
         }
       }
+      const notificationData = {
+        data: {
+          userId: resultCreateTask.rows[0].task_id,
+          eventType: "task.created",
+          title: "Новая задача",
+          body: "Создана новая задача",
+          data: {
+            task_id: resultCreateTask.rows[0].task_id,
+            task_name: data.task.task_name
+          }
+        }
+      }
+      publishMessage("task", "task.created", notificationData);
       await pool.query("COMMIT");
       return { type: "result", result: result.rows[0] };
     } catch (error) {
@@ -133,29 +160,59 @@ export const TasksModel = {
       if (sql.type == "Error") {
         throw new Error(sql.message);
       }
-      const resultUpdateTask = await pool.query(sql.message, sql.values);
+      const resultUpdateTask = await pool.query(sql.message, Array.from(sql.values.values()));
       if (!resultUpdateTask.rows.length) {
         throw new Error("The task data has not been updated");
       }
       if (data.hasOwnProperty("assignments")) {
         const task_id = resultUpdateTask.rows[0].task_id;
-        const usersId = data.assignments;
-        const checkUsersExists = await pool.query("SELECT user_id FROM users WHERE user_id = ANY($1)", [usersId]);
+        const users_id = data.assignments;
+        const checkUsersExists = await pool.query("SELECT user_id FROM users WHERE user_id = ANY($1)", [users_id]);
+        const checkUsersAssignments = await pool.query("SELECT user_id FROM task_assignments WHERE user_id = ANY($1)", [users_id]);
         const existingUsersId = checkUsersExists.rows.map(row => row.user_id);
-        const nonExistingUsers = usersId.filter(userId => !existingUsersId.includes(userId));
+        const assignmentsUsersId = checkUsersAssignments.rows.map(row => row.user_id);
+        const nonExistingUsers = users_id.filter(userId => !existingUsersId.includes(userId));
         if (nonExistingUsers.length > 0) {
           throw new Error(`Users with id ${nonExistingUsers.join(", ")} do not exist`);
         }
-        for (let userId of existingUsersId) {
+        const oldAssignmentsUsers = users_id.filter(userId => assignmentsUsersId.includes(userId));
+        const newAssignmentsUsers = users_id.filter(userId => !assignmentsUsersId.includes(userId));
+        for (const user_id of oldAssignmentsUsers) {
+          const notificationData = {
+            data: {
+              userId: user_id,
+              eventType: "task.updated",
+              title: "Обновление задачи",
+              body: "Обновлены данные задачи",
+              data: {
+                task_id: task_id,
+                task_name: data.task.task_name,
+              }
+            }
+          }
+          publishMessage("task", "task.updated", notificationData);
+        }
+        for (const user_id of newAssignmentsUsers) {
           const resultTaskAssignments = await pool.query(`INSERT INTO task_assignments (task_id, user_id, assigned_at, is_completed) \
-          VALUES ($1, $2, DEFAULT, DEFAULT) RERURNING task_assignment_id`, [task_id, userId]);
+          VALUES ($1, $2, DEFAULT, DEFAULT) RERURNING task_assignment_id`, [task_id, user_id]);
+          const notificationData = {
+            data: {
+              userId: user_id,
+              eventType: "task.assigned",
+              title: "Новая задача",
+              body: "Поручено новая задача",
+              data: {
+                task_id: task_id,
+                task_name: data.task.task_name,
+              }
+            }
+          }
+          publishMessage("task", "task.assigned", notificationData);
           if (!resultTaskAssignments.rows.length) {
             console.error("Error when insert data to task_assignment_id table", resultTaskAssignments);
             throw new Error("Error when insert data to task_assignment_id table");
           }
         }
-        await pool.query("COMMIT");
-        return { type: "result", result: `The task ${task_id} with assignments [${existingUsersId.join(", ")}] has been updated` }
       }
       await pool.query("COMMIT");
       return { type: "result", result: `The task ${taskId} has been updated` };
@@ -169,12 +226,40 @@ export const TasksModel = {
   },
   deleteTaskById: async (taskId) => {
     try {
-      const result = await pool.query("DELETE FROM tasks WHERE task_id = $1 RETURNING task_id", [taskId])
+      await pool.query("BEGIN");
+      const result = await pool.query("DELETE FROM tasks WHERE task_id = $1 RETURNING task_id, task_name, author_id", [taskId])
       if (!result.rows.length) {
-        return { type: "errorMsg", errorMsg: `Couldn't delete task ${taskId}` };
+        throw new Error(`Couldn't delete task ${taskId}`);
       }
+      const deletedDataTask = result.rows.map(row => {
+        return {
+          task_id: row.task_id,
+          task_name: row.task_name,
+          userId: row.author_id
+        }
+      });
+      for (let dataTask of deletedDataTask) {
+        const notificationData = {
+          data: {
+            userId: dataTask.userId,
+            eventType: "task.deleted",
+            title: "Удалена задача",
+            body: `Удалена задача ${dataTask.task_name}`,
+            data: {
+              task_id: dataTask.task_id,
+              task_name: dataTask.task_name
+            }
+          }
+        }
+        publishMessage("task", "task.deleted", notificationData);
+      }
+      await pool.query("COMMIT");
       return { type: "result", result: result.rows[0] };
     } catch (error) {
+      await pool.query("ROLLBACK");
+      if (error instanceof Error) {
+        return { type: "errorMsg", errorMsg: error.message };
+      }
       return { type: "errorMsg", errorMsg: "Error in Model deleteTask" };
     }
   },
